sap-bydm-transformer/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
minio==7.1.16
lxml==4.9.3
xmltodict==0.13.0
python-dotenv==1.0.0
loguru==0.7.2
pandas==2.1.1
openpyxl==3.1.2